Add COM Wrappers (#6)
* Add COM init function

* Add convenience export

* Add doc for BSTR display func

* Add `chars` method to BSTR

* Remove DWORD from public api in `Process`

* Add `into_raw` func for `BStr`

* Add missing import

+++ src/processthreadsapi.rs
@@ -34,8 +34,8 @@ impl Process {
     /// # Errors
     /// Fails if the process could not be opened.
     ///
-    pub fn open(access_rights: ProcessAccessRights, pid: DWORD) -> std::io::Result<Self> {
-        let handle = unsafe { OpenProcess(access_rights.bits(), FALSE, pid) };
+    pub fn open(access_rights: ProcessAccessRights, pid: u32) -> std::io::Result<Self> {
+        let handle = unsafe { OpenProcess(access_rights.bits(), FALSE, pid as DWORD) };
 
         if handle.is_null() {
             Err(std::io::Error::last_os_error())
